+++ cmake/nfxSiliconSources.cmake
@@ -32,6 +32,7 @@ set(NFX_SILICON_SOURCES
     src/gate/SRLatch.cpp
 
     # 74xx packages
+    src/gate/sn74/LS04.cpp
     src/gate/sn74/LS74.cpp
     src/gate/sn74/LS373.cpp
     src/gate/sn74/LS374.cpp