+++ include/nfx/Silicon.h
@@ -46,7 +46,7 @@
 #include "nfx/silicon/gate/SRLatch.h"
 
 // 74xx packages
-
+#include "nfx/silicon/gate/sn74xxx/LS04.h"
 #include "nfx/silicon/gate/sn74xxx/LS74.h"
 #include "nfx/silicon/gate/sn74xxx/LS373.h"
 #include "nfx/silicon/gate/sn74xxx/LS374.h"